requirements.txt
# requirements.txt
llama-index==0.12.52
chromadb==1.0.15
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
sentence-transformers==5.0.0
tqdm==4.67.1

python-multipart==0.0.20

llama-index-readers-file==0.4.11
llama-index-vector-stores-chroma==0.4.2
llama-index-embeddings-ollama==0.6.0
